test: isolate provider request normalization property
